+++ packages/vrender/src/jsx/jsx-classic.ts
@@ -2,6 +2,13 @@ import { isArray, isString } from '@visactor/vutils';
 import { graphicCreator } from '../graphic';
 import { REACT_TO_CANOPUS_EVENTS } from './graphicType';
 
+function flatten(list: any, out: any[]): void {
+  if (isArray(list)) {
+    return list.forEach(i => flatten(i, out));
+  }
+  out.push(list);
+}
+
 export function jsx(type: string | any, config: Record<string, any>, ...children: any) {
   const { key, attribute, ...props } = config || {};
 
@@ -10,27 +17,51 @@ export function jsx(type: string | any, config: Record<string, any>, ...children
     c = graphicCreator[type];
   }
 
-  let childrenList = [];
+  const childrenList: any[] = [];
   if (children.length) {
-    childrenList = children.length === 1 ? children[0] : children;
+    flatten(children.length === 1 ? children[0] : children, childrenList);
   }
 
+  if (c.type === 'richtext') {
+  }
   const g = c.prototype.type ? new c(attribute) : c(config);
-  if (childrenList && isArray(childrenList)) {
+  parseToGraphic(g, childrenList, props);
+
+  return g;
+}
+
+function parseToGraphic(g: any, childrenList: any[], props: any) {
+  let isGraphic: boolean = false;
+  let out: any;
+  switch (g.type) {
+    case 'richtext':
+      break;
+    case 'rich/text':
+      out = g.attribute || {};
+      childrenList[0] && (out.text = childrenList[0]);
+      break;
+    case 'rich/image':
+      break;
+    default:
+      isGraphic = true;
+  }
+
+  if (isGraphic) {
     childrenList.forEach((c: any) => {
       c && g.add(c);
     });
-  } else {
-    childrenList && g.add(childrenList);
-  }
 
-  Object.keys(props).forEach(k => {
-    const en = REACT_TO_CANOPUS_EVENTS[k];
-    if (en) {
-      g.on(en, props[k]);
+    Object.keys(props).forEach(k => {
+      const en = REACT_TO_CANOPUS_EVENTS[k];
+      if (en) {
+        g.on(en, props[k]);
+      }
+    });
+  } else {
+    if (g.type === 'richtext') {
+      g.attribute.textConfig = childrenList.map(item => item.attribute).filter(item => item);
     }
-  });
-  return g;
+  }
 }
 
 export class Fragment {